+++ CharacterManager/Heritage.sql
@@ -1,0 +1,2 @@
+UPDATE Pieces SET Discriminator = 'Piece' WHERE Discriminator IS NULL OR Discriminator = '';
+UPDATE Personnages SET Discriminator = 'Personnage' WHERE Discriminator IS NULL OR Discriminator = '';